+++ TCEngine/src/Utils/Memory.inl
@@ -1,7 +1,5 @@
 
 #include "Memory.h"
-#include <memory>
-#include <cstring>
 
 namespace Tce {
 
@@ -34,7 +32,7 @@ namespace Tce {
 
     template<typename T>
     inline void Memory::CopyArray(T *des, T *src, size_t count) {
-        for (int i = 0; i < count; i++) {
+        for (size_t i = 0; i < count; i++) {
             des[i] = src[i];
         }
     }
